# swagger_output.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EGI Media AI Backend API",
    "version": "0.1.0",
    "description": "API contract for the EGI Media AI decision-support service. The service is tenant- and company-scoped, reads editorial articles from egi-media-backend, and exposes only validated AI projections. It never exposes prompts, provider credentials, raw provider output, or CMS write operations."
  },
  "servers": [
    {
      "url": "http://localhost:5003",
      "description": "Local AI backend"
    }
  ],
...
  "paths": {
...
    "/api/v1/reports/{reportId}/review": {
      "post": {
        "tags": ["Reports"],
        "summary": "Submit a report for review or request changes",
        "operationId": "reviewReport",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/ReportIdPath" },
          { "$ref": "#/components/parameters/CompanyIdHeader" },
          { "$ref": "#/components/parameters/IdempotencyKeyHeader" },
          { "$ref": "#/components/parameters/IfMatchHeader" },
          { "$ref": "#/components/parameters/RequestIdHeader" }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReportReviewRequest" } } } },
        "responses": {
          "200": { "description": "Review state updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReportResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/BusinessRuleError" }
        }
      }
    },
    "/api/v1/reports/{reportId}/approve": {
      "post": {
        "tags": ["Reports"],
        "summary": "Approve the current report version",
        "description": "Human-only command. The model cannot approve a report. Exact approver role is product/security controlled.",
        "operationId": "approveReport",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/ReportIdPath" },
          { "$ref": "#/components/parameters/CompanyIdHeader" },
          { "$ref": "#/components/parameters/IdempotencyKeyHeader" },
          { "$ref": "#/components/parameters/IfMatchHeader" },
          { "$ref": "#/components/parameters/RequestIdHeader" }
        ],
        "responses": {
          "200": { "description": "Report approved", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReportResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/BusinessRuleError" }
        }
      }
    },
    "/api/v1/reports/{reportId}/share": {
      "post": {
        "tags": ["Reports"],
        "summary": "Create a report share delivery intent",
        "description": "External delivery command. It accepts only an approved current report, records an auditable intent, and may be held for human review. It never sends from unvalidated AI output.",
        "operationId": "shareReport",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/ReportIdPath" },
          { "$ref": "#/components/parameters/CompanyIdHeader" },
          { "$ref": "#/components/parameters/IdempotencyKeyHeader" },
          { "$ref": "#/components/parameters/IfMatchHeader" },
          { "$ref": "#/components/parameters/RequestIdHeader" }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReportShareRequest" } } } },
        "responses": {
          "202": { "description": "Share intent accepted or queued for review", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ShareIntentResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/BusinessRuleError" }
        }
      }
    },
...
    "/api/v1/internal/source/articles/{articleId}": {
      "get": {
        "tags": ["Internal Source Boundary"],
        "summary": "Read a published CMS article through the source gate",
        "description": "Read-only CMS lookup. Unpublished, deleted, malformed, or unsupported-locale articles are rejected. The citation URL is generated with the requested locale.",
        "operationId": "getPublishedSourceArticle",
        "security": [{ "internalServiceAuth": [] }],
        "parameters": [
          { "name": "articleId", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/Uuid" } },
          { "name": "locale", "in": "query", "required": false, "schema": { "type": "string", "enum": ["id", "en", "uz"], "default": "id" } },
          { "$ref": "#/components/parameters/RequestIdHeader" }
        ],
        "responses": {
          "200": { "description": "Published source article and locale-aware citation" },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "422": { "$ref": "#/components/responses/BusinessRuleError" },
          "502": { "$ref": "#/components/responses/ProviderError" },
          "503": { "$ref": "#/components/responses/ServiceUnavailable" }
        }
      }
    },
...
    "/api/v1/internal/reports/drafts": { "post": { "tags": ["Report Draft"], "summary": "Create a daily, weekly, or monthly report draft", "description": "Creates a backend-owned report draft from a period-consistent metric set and validated issue/insight pack. The input excludes raw article bodies and does not use the dashboard Top 5 as an implicit cap.", "security": [{ "internalServiceAuth": [] }], "parameters": [{ "$ref": "#/components/parameters/IdempotencyKeyHeader" }, { "$ref": "#/components/parameters/RequestIdHeader" }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReportDraftRequest" } } } }, "responses": { "200": { "description": "Report draft created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReportDraftResponse" } } } }, "400": { "$ref": "#/components/responses/ValidationError" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" } } } },
    "/api/v1/reports/{reportId}/narrative/{reportNarrativeId}/rewrite": { "post": { "tags": ["Report Review"], "summary": "Constrained rewrite of one report span", "description": "Human-only T14 rewrite. The caller selects exactly one cited span and supplies one bounded instruction. Backend preserves the existing citation set and checks the narrative version before applying the replacement.", "security": [{ "bearerAuth": [] }], "parameters": [{ "$ref": "#/components/parameters/ReportIdPath" }, { "name": "reportNarrativeId", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/Uuid" } }, { "$ref": "#/components/parameters/IdempotencyKeyHeader" }, { "$ref": "#/components/parameters/IfMatchHeader" }, { "$ref": "#/components/parameters/RequestIdHeader" }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ConstrainedRewriteRequest" } } } }, "responses": { "200": { "description": "Rewritten cited span", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ConstrainedRewriteResponse" } } } }, "400": { "$ref": "#/components/responses/ValidationError" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "409": { "$ref": "#/components/responses/Conflict" }, "422": { "$ref": "#/components/responses/BusinessRuleError" }, "502": { "$ref": "#/components/responses/ProviderError" } } } },
    "/api/v1/internal/reports/{reportId}/narrative": { "post": { "tags": ["Report Draft"], "summary": "Generate the draft report narrative", "description": "Runs T13 Mini only on the backend-created draft's validated issue and insight pack. It cannot select issues, calculate metrics, approve, share, or send the report.", "security": [{ "internalServiceAuth": [] }], "parameters": [{ "$ref": "#/components/parameters/ReportIdPath" }, { "$ref": "#/components/parameters/IdempotencyKeyHeader" }, { "$ref": "#/components/parameters/RequestIdHeader" }], "requestBody": { "required": false, "content": { "application/json": { "schema": { "type": "object", "additionalProperties": false, "maxProperties": 0 } } } }, "responses": { "200": { "description": "Draft narrative generated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReportNarrativeResponse" } } } }, "400": { "$ref": "#/components/responses/ValidationError" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "422": { "$ref": "#/components/responses/BusinessRuleError" }, "502": { "$ref": "#/components/responses/ProviderError" } } } },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT", "description": "End-user identity. Tenant/company scope is derived and verified server-side." },
      "internalServiceAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "Service JWT", "description": "Internal service or authorized operator identity; never available to frontend." }
    },
    "parameters": {
...
      "CompanyIdHeader": { "name": "X-Company-Id", "in": "header", "required": true, "description": "Active company for a mutation. Server verifies tenant/company access.", "schema": { "$ref": "#/components/schemas/Uuid" } },
...
      "ReportIdPath": { "name": "reportId", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/Uuid" } },
...
      "IdempotencyKeyHeader": { "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string", "minLength": 16, "maxLength": 255 } },
      "IfMatchHeader": { "name": "If-Match", "in": "header", "required": true, "description": "Current resource version for optimistic concurrency.", "schema": { "type": "string", "pattern": "^[0-9]+$" } },
      "RequestIdHeader": { "name": "X-Request-Id", "in": "header", "required": false, "schema": { "type": "string", "maxLength": 100 } }
    },
    "responses": {
      "Unauthorized": { "description": "Authentication missing or invalid", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "Forbidden": { "description": "Caller lacks required tenant/company/role permission", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "NotFound": { "description": "Resource not found in caller scope", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "Conflict": { "description": "Idempotency key, version, or state conflict", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "ValidationError": { "description": "Request schema validation failed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "BusinessRuleError": { "description": "Valid request cannot proceed because of pipeline/state/delivery business rules", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "ProviderError": { "description": "External AI/provider request failed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "ServiceUnavailable": { "description": "Required backend dependency is unavailable", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
    },
    "schemas": {
      "Uuid": { "type": "string", "format": "uuid" },
      "Timestamp": { "type": "string", "format": "date-time" },
...
      "Priority": { "type": "string", "enum": ["tinggi", "sedang", "rendah"] },
...
      "ReportType": { "type": "string", "enum": ["harian", "mingguan", "bulanan"] },
      "ReportReviewStatus": { "type": "string", "enum": ["draft", "in-review", "approved", "shared", "needs-review"] },
...
      "ErrorCode": { "type": "string", "enum": ["VALIDATION_ERROR", "UNAUTHORIZED", "FORBIDDEN", "NOT_FOUND", "VERSION_CONFLICT", "IDEMPOTENCY_CONFLICT", "BUSINESS_RULE_FAILED", "AI_OUTPUT_INVALID", "SOURCE_VALIDATION_FAILED", "PIPELINE_PENDING", "INTERNAL_ERROR"] },
      "ApiMeta": {
        "type": "object",
        "required": ["request_id"],
        "properties": {
          "request_id": { "type": "string" },
          "trace_id": { "type": "string" },
          "page": { "type": "integer", "minimum": 1 },
          "limit": { "type": "integer", "minimum": 1 },
          "total": { "type": "integer", "minimum": 0 },
          "total_pages": { "type": "integer", "minimum": 0 }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["success", "error", "meta"],
        "properties": {
          "success": { "type": "boolean", "enum": [false] },
          "error": { "type": "object", "required": ["code", "message"], "properties": { "code": { "$ref": "#/components/schemas/ErrorCode" }, "message": { "type": "string" }, "fields": { "type": "array", "items": { "type": "object", "properties": { "field": { "type": "string" }, "code": { "type": "string" } } } } } },
          "meta": { "$ref": "#/components/schemas/ApiMeta" }
        }
      },
      "SourceCitation": {
        "type": "object",
        "required": ["source_article_id", "locale", "url", "published_at"],
        "properties": {
          "source_article_id": { "$ref": "#/components/schemas/Uuid" },
          "locale": { "type": "string", "enum": ["id", "en", "uz"] },
          "url": { "type": "string", "format": "uri", "description": "Backend-generated canonical public URL: /{locale}/articles/{id}." },
          "title": { "type": "string", "maxLength": 255 },
          "published_at": { "$ref": "#/components/schemas/Timestamp" }
        }
      },
...
      "Report": {
        "type": "object",
        "required": ["id", "company_id", "type", "period_start", "period_end", "timezone", "review_status", "version", "updated_at"],
        "properties": {
          "id": { "$ref": "#/components/schemas/Uuid" },
          "company_id": { "$ref": "#/components/schemas/Uuid" },
          "type": { "$ref": "#/components/schemas/ReportType" },
          "period_start": { "type": "string", "format": "date" },
          "period_end": { "type": "string", "format": "date" },
          "timezone": { "type": "string", "example": "Asia/Jakarta" },
          "review_status": { "$ref": "#/components/schemas/ReportReviewStatus" },
          "version": { "type": "integer", "minimum": 1 },
          "updated_at": { "$ref": "#/components/schemas/Timestamp" },
          "summary": { "type": "string", "maxLength": 2000 }
        }
      },
      "InternalReportDraft": { "type": "object", "required": ["report_id", "report_type", "period_start", "period_end", "timezone", "context_version", "metrics", "selected_issue_pack", "review_status", "version", "created_at", "updated_at"], "properties": { "report_id": { "$ref": "#/components/schemas/Uuid" }, "report_type": { "$ref": "#/components/schemas/ReportType" }, "period_start": { "$ref": "#/components/schemas/Timestamp" }, "period_end": { "$ref": "#/components/schemas/Timestamp" }, "timezone": { "type": "string" }, "context_version": { "type": "integer", "minimum": 1 }, "metrics": { "type": "object" }, "selected_issue_pack": { "type": "array", "items": { "$ref": "#/components/schemas/ValidatedReportIssue" } }, "review_status": { "type": "string", "enum": ["draft", "needs_review"] }, "version": { "type": "integer", "minimum": 1 }, "created_at": { "$ref": "#/components/schemas/Timestamp" }, "updated_at": { "$ref": "#/components/schemas/Timestamp" } } },
      "ReportNarrative": { "type": "object", "required": ["report_narrative_id", "report_id", "prompt_version", "narrative", "review_status", "version", "created_at", "updated_at"], "properties": { "report_narrative_id": { "$ref": "#/components/schemas/Uuid" }, "report_id": { "$ref": "#/components/schemas/Uuid" }, "prompt_version": { "type": "string" }, "narrative": { "type": "object" }, "review_status": { "type": "string", "enum": ["draft"] }, "version": { "type": "integer", "minimum": 1 }, "created_at": { "$ref": "#/components/schemas/Timestamp" }, "updated_at": { "$ref": "#/components/schemas/Timestamp" } } },
...
      "ReportReviewRequest": { "type": "object", "required": ["action", "version"], "additionalProperties": false, "properties": { "action": { "type": "string", "enum": ["submit", "request_changes"] }, "version": { "type": "integer", "minimum": 1 }, "comment": { "type": "string", "maxLength": 4000 } } },
      "ReportShareRequest": { "type": "object", "required": ["version", "recipient_refs"], "additionalProperties": false, "properties": { "version": { "type": "integer", "minimum": 1 }, "recipient_refs": { "type": "array", "minItems": 1, "maxItems": 100, "items": { "type": "string", "maxLength": 255 } }, "message": { "type": "string", "maxLength": 1000 } } },
...
      "ShareIntentResponse": { "type": "object", "required": ["success", "data", "meta"], "properties": { "success": { "type": "boolean", "enum": [true] }, "data": { "type": "object", "required": ["id", "state"], "properties": { "id": { "$ref": "#/components/schemas/Uuid" }, "state": { "type": "string", "enum": ["queued", "waiting_review", "suppressed"] } } }, "meta": { "$ref": "#/components/schemas/ApiMeta" } } },
...
      "ReportResponse": { "type": "object", "required": ["success", "data", "meta"], "properties": { "success": { "type": "boolean", "enum": [true] }, "data": { "$ref": "#/components/schemas/Report" }, "meta": { "$ref": "#/components/schemas/ApiMeta" } } },
...
      "ValidatedReportIssue": { "type": "object", "required": ["report_item_id", "issue_id", "analysis_id", "priority", "title", "one_liner", "analysis", "claims", "citations"], "additionalProperties": false, "properties": { "report_item_id": { "type": "string" }, "issue_id": { "$ref": "#/components/schemas/Uuid" }, "analysis_id": { "$ref": "#/components/schemas/Uuid" }, "priority": { "$ref": "#/components/schemas/Priority" }, "title": { "type": "string", "maxLength": 180 }, "one_liner": { "type": "string", "maxLength": 320 }, "analysis": { "type": "object", "required": ["what_happened", "why_matters"], "properties": { "what_happened": { "type": "string", "maxLength": 1200 }, "why_matters": { "type": "string", "maxLength": 1200 } } }, "claims": { "type": "array", "minItems": 1, "maxItems": 12, "items": { "type": "object", "required": ["claim_id", "text", "source_article_ids"], "properties": { "claim_id": { "type": "string" }, "text": { "type": "string", "maxLength": 1200 }, "source_article_ids": { "type": "array", "minItems": 1, "items": { "type": "string" } } } } }, "citations": { "type": "array", "minItems": 1, "items": { "$ref": "#/components/schemas/SourceCitation" } } } },
      "ReportDraftRequest": { "type": "object", "required": ["report_type", "period_start", "period_end", "timezone", "context_version", "metrics", "selected_issue_pack"], "additionalProperties": false, "properties": { "report_type": { "$ref": "#/components/schemas/ReportType" }, "period_start": { "$ref": "#/components/schemas/Timestamp" }, "period_end": { "$ref": "#/components/schemas/Timestamp" }, "timezone": { "type": "string" }, "context_version": { "type": "integer", "minimum": 1 }, "metrics": { "type": "object", "required": ["period_start", "period_end"], "additionalProperties": true }, "selected_issue_pack": { "type": "array", "minItems": 1, "maxItems": 20, "items": { "$ref": "#/components/schemas/ValidatedReportIssue" } } } },
      "ReportDraftResponse": { "type": "object", "required": ["success", "data", "meta"], "properties": { "success": { "type": "boolean", "enum": [true] }, "data": { "$ref": "#/components/schemas/InternalReportDraft" }, "meta": { "$ref": "#/components/schemas/ApiMeta" } } },
      "ReportNarrativeResponse": { "type": "object", "required": ["success", "data", "meta"], "properties": { "success": { "type": "boolean", "enum": [true] }, "data": { "type": "object", "required": ["report", "narrative", "reused"], "properties": { "report": { "$ref": "#/components/schemas/InternalReportDraft" }, "narrative": { "$ref": "#/components/schemas/ReportNarrative" }, "reused": { "type": "boolean" } } }, "meta": { "$ref": "#/components/schemas/ApiMeta" } } },
      "ConstrainedRewriteRequest": { "type": "object", "required": ["allowed_span_id", "instruction", "version"], "additionalProperties": false, "properties": { "allowed_span_id": { "type": "string", "minLength": 1, "maxLength": 255 }, "instruction": { "type": "string", "minLength": 1, "maxLength": 1000 }, "version": { "type": "integer", "minimum": 1 } } },
      "ConstrainedRewriteResponse": { "type": "object", "required": ["success", "data", "meta"], "properties": { "success": { "type": "boolean", "enum": [true] }, "data": { "type": "object", "required": ["narrative", "rewritten_span", "reused"], "properties": { "narrative": { "$ref": "#/components/schemas/ReportNarrative" }, "rewritten_span": { "type": "object", "required": ["span_id", "source_claim_ids"], "properties": { "span_id": { "type": "string" }, "source_claim_ids": { "type": "array", "minItems": 1, "items": { "type": "string" } } } }, "reused": { "type": "boolean", "enum": [false] } } }, "meta": { "$ref": "#/components/schemas/ApiMeta" } } }
    }
  }
}